providers registry contracts refactor, tests refactor, token tests update
