+++ src/__tests__/providers-registry/remove-provider.spec.ts
@@ -1,0 +1,137 @@
+import ContractsTestingEnv from '../../tools/ContractsTestingEnv';
+import { ProvidersRegistryInput } from '../../providers-registry/types';
+
+const contractSrcPath = "./src/providers-registry/providers-registry.contract.ts";
+const initialState = `{
+  "trace": false,
+  "readonly": false,
+  "contractAdmins": ["bYz5YKzHH97983nS8UWtqjrlhBHekyy-kvHt_eBxBBY"]
+}`;
+
+describe('Provider Registry Contract', () => {
+  const testEnv = new ContractsTestingEnv();
+  let providersContractId: string;
+
+  beforeEach(() => {
+    providersContractId = testEnv.deployContract(contractSrcPath, JSON.parse(initialState));
+  });
+
+  afterEach(() => {
+    testEnv.clearContracts();
+  });
+
+  describe('removeProvider function', () => {
+    // adding two providers before execution of each test
+    const user1 = 'bYz5YKzHH97983nS8UWtqjrlhBHekyy-kvHt_111111';
+    const user2 = 'bYz5YKzHH97983nS8UWtqjrlhBHekyy-kvHt_222222';
+    const nonProviderAdminUser = 'bYz5YKzHH97983nS8UWtqjrlhBHekyy-kvHt_222221';
+
+    beforeEach(async () => {
+      await testEnv.interact<ProvidersRegistryInput>(
+        user1,
+        providersContractId,
+        {
+          function: 'registerProvider',
+          data: {
+            'provider': {
+              'adminsPool': [],
+              'profile': {
+                'name': 'test-provider-1',
+                'description': 'desc-1',
+                'url': 'https://test-provider-1.ok',
+              },
+            },
+          },
+        });
+
+      await testEnv.interact<ProvidersRegistryInput>(
+        user2,
+        providersContractId,
+        {
+          function: 'registerProvider',
+          data: {
+            'provider': {
+              'adminsPool': [],
+              'profile': {
+                'name': 'test-provider-2',
+                'description': 'desc-1',
+                'url': 'https://test-provider-1.ok',
+              },
+            },
+          },
+        });
+    });
+
+    it('removes provider with given provider id', async () => {
+      const interaction = await testEnv.interact<ProvidersRegistryInput>(
+        user2,
+        providersContractId,
+        {
+          function: 'removeProvider',
+          data: {
+            providerId: user2,
+          },
+        });
+
+      expect(interaction.state.providers).toEqual(
+        {
+          'bYz5YKzHH97983nS8UWtqjrlhBHekyy-kvHt_111111': {
+            'adminsPool': [user1],
+            'profile': {
+              'name': 'test-provider-1',
+              'description': 'desc-1',
+              'url': 'https://test-provider-1.ok',
+              'id': 'provider_TX-ID-1000',
+            },
+            'manifests': [],
+            'registerHeight': 1000,
+          },
+        });
+    });
+
+    it('throws if providerId not set', async () => {
+
+      const data = JSON.parse(`{}`);
+
+      await expect(testEnv.interact<ProvidersRegistryInput>(
+        user2,
+        providersContractId,
+        {
+          function: 'removeProvider',
+          data,
+        }))
+        .rejects
+        .toThrowError('\'providerId\' field is required.');
+    });
+
+    it('throws if provider does not exist', async () => {
+      await expect(testEnv.interact<ProvidersRegistryInput>(
+        user2,
+        providersContractId,
+        {
+          function: 'removeProvider',
+          data: {
+            providerId: 'bYz5YKzHH97983nS8UWtqjrlhBHekyy-kvHt_333333',
+          },
+        }))
+        .rejects
+        .toThrowError('Provider with id bYz5YKzHH97983nS8UWtqjrlhBHekyy-kvHt_333333 is not registered.');
+    });
+
+    it('throws if caller has no privileges to provider ', async () => {
+      await expect(testEnv.interact<ProvidersRegistryInput>(
+        nonProviderAdminUser,
+        providersContractId,
+        {
+          function: 'removeProvider',
+          data: {
+            providerId: user2,
+          },
+        }))
+        .rejects
+        .toThrowError('bYz5YKzHH97983nS8UWtqjrlhBHekyy-kvHt_222221 is not an admin for bYz5YKzHH97983nS8UWtqjrlhBHekyy-kvHt_222222');
+    });
+  });
+
+
+});
