+++ src/providers-registry/actions/write/removeProvider.ts
@@ -1,0 +1,16 @@
+import { ProvidersRegistryAction, ProvidersRegistryState, RemoveProviderData } from '../../types';
+import { checkPrivileges, checkProviderExists, checkProviderId } from '../_commons';
+
+export const removeProvider = async (
+  state: ProvidersRegistryState,
+  { caller, input: { data } }: ProvidersRegistryAction
+) => {
+  const removeProviderData = data as RemoveProviderData;
+  checkProviderId(removeProviderData.providerId);
+  checkProviderExists(removeProviderData.providerId, state.providers);
+  checkPrivileges(caller, removeProviderData.providerId, state.providers);
+
+  delete state.providers[removeProviderData.providerId];
+
+  return { state };
+}
