+++ src/providers-registry/actions/write/switchReadonly.ts
@@ -1,0 +1,10 @@
+import { ProvidersRegistryAction, ProvidersRegistryState } from '../../types';
+
+export const switchReadonly = async (
+  state: ProvidersRegistryState,
+  { caller, input: { data } }: ProvidersRegistryAction,
+) => {
+  state.readonly = !state.readonly;
+
+  return { state };
+};
