+++ src/providers-registry/actions/write/switchTrace.ts
@@ -1,0 +1,9 @@
+import { ProvidersRegistryAction, ProvidersRegistryState } from '../../types';
+
+export const switchTrace = async (
+  state: ProvidersRegistryState,
+  { caller, input: { data } }: ProvidersRegistryAction,
+) => {
+  state.trace = !state.trace;
+  return { state };
+};
